ML/requirements.txt
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
catboost==1.2.2
psycopg2-binary==2.9.9
joblib==1.3.2
matplotlib==3.8.0
seaborn==0.13.0
python-dotenv==1.0.0